backups/2025-01-29T14-26-12-685Z/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1